refactor(portal-next): move application log table to separate component
