+++ gravitee-apim-portal-webui-next/src/app/applications/application/application-tab-logs/application-log-table/application-log-table.component.html
@@ -188,8 +188,18 @@
         <td mat-cell *matCellDef="let log" [class.error]="log.status >= 300">{{ log.status }}</td>
       </ng-container>
 
+      <ng-container matColumnDef="action">
+        <th mat-header-cell *matHeaderCellDef></th>
+        <td mat-cell [routerLink]="['./', element.id]" *matCellDef="let element">
+          <mat-icon>arrow_right</mat-icon>
+        </td>      </ng-container>
+
       <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row class="application-logs__table-row" *matRowDef="let application; columns: displayedColumns"></tr>
+      <tr
+        mat-row
+        class="application-logs__table-row"
+        *matRowDef="let log; columns: displayedColumns"
+        [routerLink]="['./', log.id]"></tr>
     </table>
     <div class="application-logs__table__pagination">
       <button mat-icon-button [disabled]="!pagination().hasPreviousPage" (click)="goToPreviousPage()" aria-label="Previous page of logs">
