+++ gravitee-apim-portal-webui-next/src/app/applications/application/application-tab-logs/application-log-table/application-log-table.component.scss
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@use '../../../../scss/theme';
+@use '../../../../../scss/theme';
 
 $form-field-hint-height: 16px;
 
@@ -37,6 +37,7 @@ $form-field-hint-height: 16px;
 
     &-row {
       background: var(--mdc-outlined-card-container-color);
+      cursor: pointer;
     }
 
     &__pagination {
